cleanup: move code
