+++ src/base/decode.h
@@ -11,8 +11,8 @@ void decode_reset(VGMSTREAM* vgmstream);
  * buffer already, and we have samples_to_do consecutive samples ahead of us. */
 void decode_vgmstream(VGMSTREAM* vgmstream, int samples_written, int samples_to_do, sample_t* buffer);
 
-/* Detect loop start and save values, or detect loop end and restore (loop back). Returns 1 if loop was done. */
-int decode_do_loop(VGMSTREAM* vgmstream);
+/* Detect loop start and save values, or detect loop end and restore (loop back). Returns true if loop was done. */
+bool decode_do_loop(VGMSTREAM* vgmstream);
 
 /* Calculate number of consecutive samples to do (taking into account stopping for loop start and end) */
 int decode_get_samples_to_do(int samples_this_block, int samples_per_frame, VGMSTREAM* vgmstream);
