+++ src/base/info.c
@@ -6,6 +6,8 @@
 #include "../util/channel_mappings.h"
 #include "../util/sf_utils.h"
 
+#define TEMPSIZE (256+32)
+
 /*******************************************************************************/
 /* TEXT                                                                        */
 /*******************************************************************************/
@@ -21,7 +23,6 @@ static void describe_get_time(int32_t samples, int sample_rate, double* p_time_m
 /* Write a description of the stream into array pointed by desc, which must be length bytes long.
  * Will always be null-terminated if length > 0 */
 void describe_vgmstream(VGMSTREAM* vgmstream, char* desc, int length) {
-#define TEMPSIZE (256+32)
     char temp[TEMPSIZE];
     double time_mm, time_ss;
