+++ src/libvgmstream.vcxproj.filters
@@ -487,9 +487,6 @@
     <ClCompile Include="base\api_tags.c">
       <Filter>base\Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="base\config.c">
-      <Filter>base\Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="base\decode.c">
       <Filter>base\Source Files</Filter>
     </ClCompile>
@@ -512,6 +509,12 @@
       <Filter>base\Source Files</Filter>
     </ClCompile>
     <ClCompile Include="base\mixing_macros.c">
+      <Filter>base\Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="base\play_config.c">
+      <Filter>base\Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="base\play_state.c">
       <Filter>base\Source Files</Filter>
     </ClCompile>
     <ClCompile Include="base\plugins.c">
@@ -1718,9 +1721,6 @@
       <Filter>meta\Source Files</Filter>
     </ClCompile>
     <ClCompile Include="meta\ps2_wmus.c">
-      <Filter>meta\Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="meta\ps2_xa30.c">
       <Filter>meta\Source Files</Filter>
     </ClCompile>
     <ClCompile Include="meta\psb.c">
